force MI 2 to use wlan0 to start wifi direct

+++ manager/comp_wifi/wifi.py
@@ -439,13 +439,14 @@ def reset_wifi_interface():
 def start_hotspot_on_wcnss(ssid, password):
     control_socket_dir = get_wpa_supplicant_control_socket_dir()
     load_p2p_firmware(control_socket_dir)
-    if 'p2p0' in list_wifi_ifaces():
+    if 'p2p0' in list_wifi_ifaces() and 'MI 2' != shell_execute('getprop ro.product.model').strip():
         LOGGER.info('start p2p persistent group using p2p0')
         shell_execute('netcfg p2p0 up')
         p2p_control_socket_dir = get_p2p_supplicant_control_socket_dir()
         delete_existing_p2p_persistent_networks('p2p0', p2p_control_socket_dir)
         start_p2p_persistent_network('p2p0', p2p_control_socket_dir, ssid, password, sets_channel=True)
     else:
+        LOGGER.info('start p2p persistent group using %s' % WIFI_INTERFACE)
         delete_existing_p2p_persistent_networks(WIFI_INTERFACE, control_socket_dir)
         start_p2p_persistent_network(WIFI_INTERFACE, control_socket_dir, ssid, password, sets_channel=True)
     log_upstream_wifi_status('after p2p persistent group created', control_socket_dir)
